changed install.sh functionality and added unrar_to_dir.sh
-unrar_to_dir.sh is a script that looks for .rar files in the current
directory, then creates a directory with the .rar files name (Example:
creates my_file/ for my_file.rar) then moves the .rar file in the new
directory that was created

-install.sh now finds out which OS is being used, and calls a script
that is responsible for install the require programs using the correct
methods. Currently only Ubuntu is being supported, but this will make
it easier to support different OS's in the future. All that is needed
is to create a check for the OS, and add a script just for it.