changed install options and added fish as a availible shell
Before only zsh was used as a shell. Now when installing on an ubuntu
machine you have the option of selecting if you want zsh or fish. The
default is zsh. Because there are now two options all logic pertaining
to shells was removed from ubuntu.sh (this is where zsh was originally
installed) and added to their own seperate scripts (install_zsh.sh
and install_fish.sh) this was done so unneeded downloading will not
take place (why download zsh and all its components if you are not
going to use it). Along with adding a fish shell download option, there
are also common configs (in the form of functions) that can be used with
it.